Move ToggleSkinMode in weight_painting module
